Create session storage for gender scanning
