+++ html/ai-training-camera.html
@@ -1,0 +1,148 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Avaturn Camera Scan</title>
+    <style>
+        body {
+            margin: 0;
+            padding: 0;
+        }
+      video {
+        height: 932px;
+        width: 430px;
+      }
+      .fixed-bottom {
+    position: fixed;
+    bottom: 0;
+}
+
+footer {
+    position: fixed;
+    display: flex;
+}
+
+#home {
+    width: 60px;
+    height: 60px;
+    position: absolute;
+    left: 90px;
+    bottom: 30px;
+    background-color: black;
+    border-radius: 20px;
+}
+
+#home-icon {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+    top: 15px;
+}
+
+#camera {
+    width: 80px;
+    height: 80px;
+    position: absolute;
+    left: 175px;
+    transform: translateY(-50%);
+    bottom: 30px;
+    background-color: black;
+    border-radius: 20px;
+}
+
+#camera-icon {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+    top: 20px;
+}
+
+#bookmark {
+    width: 60px;
+    height: 60px;
+    position: absolute;
+    left: 280px;
+    bottom: 30px;
+    background-color: black;
+    border-radius: 20px;
+}
+
+#bookmark-icon {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+    top: 15px;
+}
+
+#styling-bar {
+    background-color: black;
+    height: 3px;
+    width: 50%;
+    position: absolute;
+    bottom: 10px;
+    left: 50px;
+}
+
+#content {
+    margin-bottom: 200px;
+}
+
+
+
+    </style>
+</head>
+
+<body>
+    <video id="camera-feed" autoplay muted></video>
+    <footer id="footer" class="fixed-bottom">
+        <div id="home">
+          <svg id="home-icon" width="32" height="32" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+              d="M4 10.22C4 9.43883 4.3103 8.6897 4.86264 8.13736V8.13736C6.27285 6.72715 7.9548 5.61808 9.80648 4.87741L10.5978 4.5609C11.4979 4.20083 12.5021 4.20083 13.4022 4.5609L14.1935 4.87741C16.0452 5.61808 17.7272 6.72715 19.1374 8.13736V8.13736C19.6897 8.6897 20 9.43883 20 10.22V17.0131C20 18.7195 18.7294 20.1588 17.0361 20.3705V20.3705C13.6917 20.7885 10.3083 20.7885 6.96387 20.3705V20.3705C5.27065 20.1588 4 18.7195 4 17.0131V10.22Z"
+              stroke="white" stroke-width="2" stroke-linejoin="round" />
+            <path
+              d="M9 21V16.3246C9 15.5335 9.50616 14.8313 10.2566 14.5811V14.5811C11.3883 14.2039 12.6117 14.2039 13.7434 14.5811V14.5811C14.4938 14.8313 15 15.5335 15 16.3246V21"
+              stroke="white" stroke-width="2" stroke-linejoin="round" />
+          </svg>
+        </div>
+        <div id="camera">
+            <svg id="camera-icon" width="32" height="32"   viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                <path d="M3.24258 8.65098C3.6909 7.08183 5.12512 6 6.75706 6H7.31995C7.73673 6 8.11774 5.76452 8.30413 5.39175L8.75736 4.48528C8.91272 4.17456 9.18515 3.93828 9.51472 3.82843V3.82843C11.1279 3.29069 12.8721 3.29069 14.4853 3.82843V3.82843C14.8149 3.93828 15.0873 4.17456 15.2426 4.48528L15.6959 5.39175C15.8823 5.76452 16.2633 6 16.68 6H17.2429C18.8749 6 20.3091 7.08183 20.7574 8.65098L20.7964 8.78749C21.5805 11.5317 21.5171 14.4487 20.6146 17.1563V17.1563C20.2318 18.3046 19.2889 19.1778 18.1145 19.4714L17.8567 19.5358C14.0114 20.4972 9.9886 20.4972 6.14333 19.5358L5.88545 19.4714C4.71113 19.1778 3.7682 18.3046 3.38542 17.1563V17.1563C2.48291 14.4487 2.41952 11.5317 3.20357 8.78749L3.24258 8.65098Z" stroke="white" stroke-width="2" stroke-linejoin="round"/>
+                <path d="M8.72727 10.8182C8.89896 10.3031 9.30312 9.89896 9.81818 9.72727V9.72727C11.2344 9.25519 12.7656 9.25519 14.1818 9.72727V9.72727C14.6969 9.89896 15.101 10.3031 15.2727 10.8182V10.8182C15.7448 12.2344 15.7448 13.7656 15.2727 15.1818V15.1818C15.101 15.6969 14.6969 16.101 14.1818 16.2727V16.2727C12.7656 16.7448 11.2344 16.7448 9.81818 16.2727V16.2727C9.30312 16.101 8.89896 15.6969 8.72727 15.1818V15.1818C8.25519 13.7656 8.25519 12.2344 8.72727 10.8182V10.8182Z" stroke="white" stroke-width="2" stroke-linejoin="round"/>
+                </svg>                               
+        </div>
+        <div id="bookmark">
+          <svg id="bookmark-icon" width="32" height="32" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+              d="M5.57671 7.38629C5.82785 5.37722 7.36905 3.77182 9.36621 3.43897V3.43897C11.11 3.14833 12.89 3.14833 14.6338 3.43897V3.43897C16.6309 3.77182 18.1722 5.37722 18.4233 7.38629L18.4985 7.9881C18.8315 10.6523 18.8315 13.3477 18.4985 16.0119L18.2482 18.0148C18.1198 19.0418 17.0839 19.6946 16.102 19.3673L16.0525 19.3508C15.3604 19.1201 14.7314 18.7314 14.2155 18.2155V18.2155C12.9919 16.9919 11.0081 16.9919 9.78451 18.2155V18.2155C9.26859 18.7314 8.63964 19.1201 7.94747 19.3508L7.89803 19.3673C6.91611 19.6946 5.88022 19.0418 5.75184 18.0148L5.50149 16.0119C5.16846 13.3477 5.16846 10.6523 5.50149 7.9881L5.57671 7.38629Z"
+              stroke="white" stroke-width="2" stroke-linejoin="round" />
+            <path
+              d="M10.6 7.8C10.6433 7.71345 10.7135 7.64327 10.8 7.6V7.6C11.5554 7.22229 12.4446 7.22229 13.2 7.6V7.6C13.2865 7.64327 13.3567 7.71345 13.4 7.8V7.8C13.7777 8.55542 13.7777 9.44458 13.4 10.2V10.2C13.3567 10.2865 13.2865 10.3567 13.2 10.4V10.4C12.4446 10.7777 11.5554 10.7777 10.8 10.4V10.4C10.7135 10.3567 10.6433 10.2865 10.6 10.2V10.2C10.2223 9.44458 10.2223 8.55542 10.6 7.8V7.8Z"
+              stroke="white" stroke-width="2" stroke-linejoin="round" />
+          </svg>
+        </div>
+        <div id="styling-bar"></div>
+      </footer>
+    <canvas id="canvas" style="display: none;"></canvas>
+
+    <script type="module" src="/scripts/ai-training-camera.js"></script>
+    <script>
+        // Check the selected video state on page load
+      const selectedVideo = sessionStorage.getItem('selectedVideo');
+      document.getElementById('camera').addEventListener('click', () => {
+        console.log('clicked');
+      if (selectedVideo) {
+        // Redirect to the appropriate scanning page
+        if (selectedVideo === 'female') {
+          window.location.href = "ai-training-female-body-scan.html";
+        } else if (selectedVideo === 'male') {
+          window.location.href = "ai-training-male-body-scan.html";
+        }
+    }
+});
+    </script>
+</body>
+
+</html>
