+++ html/ai-training-female-body-scan.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Welcome to FitUp</title>
+    <meta name="Comp2800 template" content="FitUp">
+
+    <!-- Required meta tags -->
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <link rel="stylesheet" type="text/css" href="/styles/body-scan.css" />
+    <style>
+        video {
+            width: 100vw;
+            height: 100vh;
+            margin: 0;
+            padding: 0;
+        }
+    </style>
+</head>
+
+<body>
+
+    <!-- <main>
+        <div id="container3D"></div>
+    </main> -->
+    <video loop="" autoplay="" muted="" playsinline="">
+        <source src="/models/female-body-scan.mp4" type="video/mp4">
+    </video>   
+    <!-- <script type="module" src="/scripts/body-scan.js"></script>  -->
+</body>
+
+</html>
