+++ html/ai-training-find-coach.html
@@ -8,6 +8,7 @@
    
     <link rel="stylesheet" href="/styles/ai-training-find-coach.css"></link>
 
+    
 </head>
 
 <body>
