+++ html/ai-training-scan-request.html
@@ -21,10 +21,11 @@
     #female {
       margin-left: 2px;
     }
-   body {
-    background-color: #606060;
-    overflow: hidden;
-   }
+
+    body {
+      background-color: #606060;
+      overflow: hidden;
+    }
   </style>
 </head>
 
@@ -110,33 +111,41 @@
   <script>
 
 document.addEventListener('DOMContentLoaded', () => {
-    const toggleVideo = (event) => {
-      const video = event.target;
-      let otherVideo = null;
+  const toggleVideo = (event) => {
+    const video = event.target;
+    let otherVideo = null;
 
+    if (video.id === 'female') {
+      otherVideo = document.getElementById('male');
+    } else if (video.id === 'male') {
+      otherVideo = document.getElementById('female');
+    }
+
+    if (otherVideo !== null) {
+      video.classList.add('fade');
+      setTimeout(() => {
+        otherVideo.classList.remove('fade');
+      }, 0);
+
+      // Save the state of which video was toggled
       if (video.id === 'female') {
-        otherVideo = document.getElementById('male');
+        sessionStorage.setItem('selectedVideo', 'female');
       } else if (video.id === 'male') {
-        otherVideo = document.getElementById('female');
+        sessionStorage.setItem('selectedVideo', 'male');
       }
+    }
+  };
 
-      if (otherVideo !== null) {
-        video.classList.add('fade');
-        setTimeout(() => {
-          otherVideo.classList.remove('fade');
-        }, 0);
-      }
-    };
-
-    ['female', 'male'].forEach((videoId) => {
-      document.getElementById(videoId).addEventListener('click', toggleVideo);
-    });
+  ['female', 'male'].forEach((videoId) => {
+    document.getElementById(videoId).addEventListener('click', toggleVideo);
   });
+});
 
     const nextPage = () => {
       // Navigate to another page
-      window.location.href = "ai-training-body-scan.html";
+      window.location.href = "ai-training-camera.html";
     };
+
 
     document.getElementById('continue').addEventListener('click', nextPage);
   </script>
