+++ html/aichat-log.html
@@ -79,6 +79,7 @@
     <form id="chat-input-form">
       <input id="chat-question" type="text" name="chat-question" placeholder="Type to start chatting">
       <div id="microphone">
+        <img id="active-microphone" src="/img/microphone.gif" alt="act"/>
         <svg id="microphone-icon" width="32" height="32" viewBox="0 0 24 24" fill="none"
           xmlns="http://www.w3.org/2000/svg">
           <path
